break out matrix features of character sheet
